Address ueshin's comments
